feat: Obtain the API key from the file it is held in
Since the API key cannot be held on a public repo (Github) you have placed
it in a file that is ignored by .gitignore.  The file holding the API key
string should be placed in the main/assets directory.  It should be the only
thing in the file, with no other characters.  Once it is in place, it will
be read and used in the project.