feat: Increase stability of app when internet has gone.
App should not crash when internet is gone.  But any pictures that have
not downloaded will not be available.  Any data that is not present,
like if the user tries to change types after connection is lost, cannot
be fetched.  If there is data, user can still navigate.  If there is no
data clicks will not go anywhere until internet is reconnected and a
query event is launced.  This can happen because of rotation, or trying
to change type after internet is restored.  A Toast indicates to the user
that the connection was lost and that it will have to be reestablished for
normal operation and update data.