feat: Begin modifying Grid Item for navigation without internet
The app should be navigabe without internet connection.  But in that event, if
there is any information available you will not have pictures, since they are
always downloaded.  So, you are preparing to be able to navigate to details
through Grid Items identified through Titles.  So even if the picture is not
there; if there is enough information left for navigation, you can do it.