feat: Add a ProgressBar so user can't see changeover from Popular to Top Rated
There was always a pause when user went from one type to another.  The user
could see the previous list for a second, even though they changed it.  So
now when the user comes back from Settings there is a ProgressBar until the
data is back from the internet, or whereever and is ready to be seen.