feat: Modify code to work if there is a Movie with both types
If there is a movie in the DB that is both Top Rated and Popular deal with that
issue effectively.  The other insert should be an update, and the any delete
of that title should actually be an update to remove the current type tag.
Once a title only has one tag, it is normal and will be treated normally.