feat: Do first attempt to fix snackbar so page loads on return from Settings
Your idea is to use a boolean that will be triggered on when the Snackbar starts
and turned off in onResume.  May not fire because internet not ready yet, or because
some other timing issue.  And may make all rotations slower for something that
hopefully won't happen that often(loss of internet before 1st use).