feat: Make Snackbar indefinite so user can't miss it
Can be swiped away.  User still has to rotate screen to restart loader.
Last thing almost worked but if the user came back before connection was
regained then we were in the same predicament.  Thought it was better for
now to just have the user in this rare circumstance, figure it out.  The
rare circumstance is just having installed the app and then lost connection
before the 1 use.  If the app had been open previously; then the user can
peruse the app without posters just fine, for the most part.