feat: Prep MovieAdapter and Main for Diaolgs, Fix Back Btn on Details
Do not use the Dialogs yet, you are now not sure if you need them here.
May want to use them on Details only.  Fix the Back Button to work like
the Up button since we know that the previous Activity for Details will
always (for now) be MainActivity.  May want to refine when the previous
Activity has to do a whole reload at a later date.  Probably should be
based on what has visibly changed.