feat: Add Dialog Choice for user when removing movie from Favorites
If they choose to remove movie, will disable the Heart so it cannot
be re-chosen.  This is because in certain circumstances (If a movie only exists
in Favorites) the title will actually be deleted and cannot be re-added to list.
Dialog added because of this and the fact that the decision is then, at times,
irreversible.  May want to ask Carlos about coloring later.

Will use:
https://guides.codepath.com/android/using-dialogfragment#styling-custom-dialog
https://stackoverflow.com/questions/32784009/styling-custom-dialog-fragment-not
-working/32791569#32791569
https://stackoverflow.com/questions/2795300/how-to-implement-a-custom-alertdialog
-view/28801551#28801551
https://stackoverflow.com/questions/4719051/custom-alertdialog-problem-with-
android-r-id-custom
https://stackoverflow.com/questions/10313382/how-to-get-elementsfindviewbyid-
for-a-layout-which-is-dynamically-loadedsetvi/18773261#18773261
Not the correct type, but close.https://developer.android.com/reference/android/
app/AlertDialog.html
https://developer.android.com/reference/android/app/DialogFragment.html#Lifecycle
https://stackoverflow.com/questions/15151783/stackoverflowerror-when-trying-to-
inflate-a-custom-layout-for-an-alertdialog-ins
https://developer.android.com/guide/topics/ui/dialogs.html
https://stuff.mit.edu/afs/sipb/project/android/docs/guide/topics/ui/dialogs.html
https://stackoverflow.com/questions/2115758/how-do-i-display-an-alert-dialog
-on-android