fix: Disappearing Views when turning a disabled Favorite Movie
Use internal State class to hold values for Details Page when on Favorites
and the Heart is disabled.  Meaning that the Movie has been deleted from the
database but the user is still kicking around the page.  Otherwise, most of
the Views that changed per movie would be missing, as well as, the Poster.
Hopefully, this will handle the case that the Favorite the User is on was
only a Favorite and was immediately deleted.