feat: Change the way the Review/Youtube data is input for better input results
You are getting FileNotFoundExceptions from the internet.  Most likely because
you are hitting the X-Rate-Limit on the server.  You are only allowed a certain
number of requests per 10 second period.  With you making requests for Movies,
Runtimes, Reviews, and Youtube videos, you are most likely hitting the limit.
So change input so User will make request to Internet for Reviews/Videos only
when they go to a specific movie.  This should also have the side-effect of
making the app faster.

You should also fix that when the User deletes a Favorite from Details the
Media Title's color goes back to the BG for popular.  I believe you are not
restoring this and the empty text values in your onRestoreInstanceState().