fix: Fix Errors that occur when there is no internet
Found a couple bugs having to do with the database and no internet.  So you should fix
them at this point.  There is also a logic error having to do with the Database and
updated items also being inserted if they were of a different type.  This caused not
all of the movies to appear in the Main page.  So you should fix that as well.